move test to test_simulation_with_experiment
