+++ tests/unit/test_simulation.py
@@ -370,70 +370,6 @@ class TestSimulation(unittest.TestCase):
             decimal=5,
         )
 
-    def test_solve_with_sensitivities_and_experiment(self):
-        model = pybamm.lithium_ion.SPM()
-        param = model.default_parameter_values
-        input_param_name = "Negative electrode active material volume fraction"
-        input_param_value = 0.6
-        param.update({input_param_name: "[input]"})
-        exp = pybamm.Experiment(
-            [
-                "Discharge at C/20 for 1 hour",
-                "Charge at 1 A until 4.1 V",
-                "Hold at 4.1 V until C/2",
-                "Discharge at 2 W for 1 hour",
-            ]
-        )
-
-        solver = pybamm.IDAKLUSolver()
-        sim = pybamm.Simulation(
-            model, parameter_values=param, experiment=exp, solver=solver
-        )
-
-        sol_sens = sim.solve(
-            inputs={input_param_name: input_param_value},
-            calculate_sensitivities=True,
-        )
-
-        sol_sens_time = sol_sens.t
-        sol_sens_all = sol_sens["Terminal voltage [V]"].sensitivities["all"]
-        print("XXXXXXXXXXXXXXXXsol_sens_all", sol_sens_all[0])
-        print("XXXXXXXXXXXXXXXXsol_sens_all", sol_sens_all[1])
-        sol_sens = (
-            sol_sens["Terminal voltage [V]"]
-            .sensitivities[input_param_name]
-            .full()
-            .flatten()
-        )
-
-        solver = pybamm.IDAKLUSolver()
-        sim = pybamm.Simulation(
-            model, parameter_values=param, experiment=exp, solver=solver
-        )
-
-        h = 1e-5
-        sol_neg = sim.solve(
-            inputs={input_param_name: input_param_value - h},
-        )
-        sol_neg = sol_neg["Terminal voltage [V]"](sol_sens_time)
-        sol_pos = sim.solve(
-            inputs={input_param_name: input_param_value + h},
-        )
-        sol_pos = sol_pos["Terminal voltage [V]"](sol_sens_time)
-
-        sol_fd = (sol_pos - sol_neg) / (2 * h)
-
-        import matplotlib.pyplot as plt
-
-        fig = plt.figure()
-        ax = fig.add_subplot(111)
-        ax.plot(sol_sens_time, sol_fd, label="FD")
-        ax.plot(sol_sens_time, sol_sens, label="Sens")
-        plt.legend()
-        plt.savefig("sensitivities.png")
-
-        np.testing.assert_array_almost_equal(sol_fd, sol_sens, decimal=5)
-
     def test_step_with_inputs(self):
         dt = 0.001
         model = pybamm.lithium_ion.SPM()
